requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20200923.3=h046ec9c_0
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py38h96a0964_0
aiosignal=1.2.0=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.2=py38hecd8cb5_1001
argcomplete=1.12.3=pyhd8ed1ab_2
argon2-cffi=20.1.0=py38h9ed2024_1
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.1=pyhd8ed1ab_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.0=py_2
backports.zoneinfo=0.2.1=py38h96a0964_4
beautifulsoup4=4.10.0=pyha770c72_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blinker=1.4=py_1
bottleneck=1.3.2=py38hf1fa96c_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py38h96a0964_1003
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2021.10.8=h033912b_0
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=h0ab9d94_1001
certifi=2021.10.8=py38h50d1736_1
cffi=1.14.6=py38h2125817_0
chardet=4.0.0=py38h50d1736_1
charset-normalizer=2.0.7=pypi_0
click=8.0.3=py38h50d1736_1
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
compressed-rtf=1.0.6=pyh9f0ad1d_0
cryptography=36.0.0=py38h56c4533_0
cycler=0.10.0=py38_0
cython=0.29.23=py38ha048514_1
cytoolz=0.11.2=py38h96a0964_1
dask-core=2021.11.2=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=h18a8e69_0
debugpy=1.4.1=py38h23ab428_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dnspython=1.16.0=py38_0
docx2txt=0.8=py_0
ebcdic=1.1.1=pyhd8ed1ab_0
ebooklib=0.17=py_0
entrypoints=0.3=py38_0
enum34=1.1.10=py38h32f6830_2
expat=2.4.1=h23ab428_2
ffmpeg=4.3.2=h4dad6da_0
fontconfig=2.13.1=h10f422b_1005
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=ha233b18_0
frozenlist=1.2.0=py38h96a0964_1
fsspec=2021.11.0=pyhd8ed1ab_0
future=0.18.2=py38h50d1736_4
gast=0.3.3=py_0
gensim=4.0.1=py38h23ab428_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=hbcb3906_2
glib=2.69.1=hdf23fa2_0
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.14=h38d11af_0
grpc-cpp=1.36.4=h33525da_1
grpcio=1.36.1=py38ha263829_0
h5py=2.10.0=nompi_py38h9a16e60_106
harfbuzz=2.8.1=ha8c618c_0
hdf5=1.10.6=hdbbcd12_0
icu=58.2=h0a44026_3
idna=3.3=pypi_0
imagecodecs-lite=2019.12.3=py38hbe852b5_4
imageio=2.9.0=pyhd3eb1b0_0
imapclient=2.2.0=pyhd8ed1ab_0
imbalanced-learn=0.8.1=pyhd8ed1ab_0
importlib-metadata=4.8.1=py38hecd8cb5_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=hecd8cb5_3375
ipykernel=6.4.1=py38hecd8cb5_1
ipython=7.27.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jasper=1.900.1=h636a363_1006
jedi=0.18.0=py38hecd8cb5_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.8.1=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.4.3=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
keras-tuner=1.0.3=pyhd8ed1ab_0
kiwisolver=1.3.1=py38h23ab428_0
krb5=1.19.2=hcfbf3a7_3
lame=3.100=h35c211d_1001
lcms2=2.12=hf1fd2bf_0
libblas=3.9.0=11_osx64_mkl
libcblas=3.9.0=11_osx64_mkl
libcurl=7.79.1=hf45b732_1
libcxx=12.0.0=h2f01273_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
liblapack=3.9.0=11_osx64_mkl
liblapacke=3.9.0=11_osx64_mkl
libnghttp2=1.43.0=h6f36284_1
libopencv=4.5.0=py38_3
libpng=1.6.37=ha441bb4_0
libprotobuf=3.15.8=hcf210ce_1
libsodium=1.0.18=h1de35cc_0
libssh2=1.10.0=h52ee1ee_2
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
libxslt=1.1.34=h83b36ba_0
llvm-openmp=12.0.0=h0dcd299_1
locket=0.2.0=py_2
lxml=4.6.4=py38h6ea4786_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=py38h9ed2024_0
matplotlib=3.4.2=py38hecd8cb5_0
matplotlib-base=3.4.2=py38h8b3ea08_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h1de35cc_1001
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.0=py38h4a7008c_2
mkl_random=1.2.2=py38hb2f4e1b_0
msg-extractor=0.28.7=pyhd8ed1ab_1
multidict=5.2.0=py38h96a0964_1
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.6=hedd7734_0
networkx=2.6.3=pyhd8ed1ab_1
nltk=3.6.5=pyhd8ed1ab_0
notebook=6.4.3=py38hecd8cb5_0
numexpr=2.7.3=py38h5873af2_1
numpy=1.20.3=py38h4b4dc7a_0
numpy-base=1.20.3=py38he0bd621_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.0=py38_3
openh264=2.1.1=hfd3ada9_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1l=h0d85af4_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.3=py38h5008ddb_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd8ed1ab_0
pathlib=1.0.1=py38h50d1736_5
pcre=8.45=h23ab428_0
pdfminer.six=20211012=pyhd8ed1ab_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38ha4cf6ea_0
pip=21.2.4=py38hecd8cb5_0
pixman=0.38.0=h01d97ff_1003
pooch=1.5.2=pyhd8ed1ab_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.15.8=py38ha048514_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=4.5.0=py38h9bb44b7_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=1.7.1=py_0
pymongo=3.12.0=py38h23ab428_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.17.3=py38haf1e3a3_0
pysocks=1.7.1=py38h50d1736_4
python=3.8.11=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd8ed1ab_1
python-pptx=0.6.21=pyhd8ed1ab_0
python-tzdata=2021.5=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pytz-deprecation-shim=0.1.0.post0=py38h50d1736_1
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.2.0=py38hbe852b5_0
pyyaml=6.0=py38h96a0964_3
pyzmq=22.2.1=py38h23ab428_1
qt=5.9.7=h468cd18_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
re2=2021.04.01=he49afe7_0
readline=8.1=h9ed2024_0
regex=2021.11.10=py38h96a0964_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.8=pyhd8ed1ab_0
scikit-image=0.18.3=py38ha53d530_0
scikit-learn=0.24.2=py38hb2f4e1b_0
scipy=1.7.1=py38h88652d9_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38hecd8cb5_0
sip=4.19.8=py38h0a44026_0
six=1.16.0=pyhd3eb1b0_0
smart_open=5.2.1=pyhd8ed1ab_0
snappy=1.1.8=hb1e8313_3
soupsieve=2.3=pyhd8ed1ab_0
speechrecognition=3.7.1=py38h50d1736_3
sqlite=3.36.0=hce871da_0
tabulate=0.8.9=pyhd8ed1ab_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-data-server=0.6.0=py38h56c4533_1
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow=2.4.0=py38h50d1736_0
tensorflow-base=2.4.0=py38h428766a_0
tensorflow-estimator=2.4.0=pyh9656e83_0
termcolor=1.1.0=py_2
terminado=0.9.4=py38hecd8cb5_0
terminaltables=3.1.0=py_0
testpath=0.5.0=pyhd3eb1b0_0
textract=1.6.4=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2019.7.26.2=py38_0
tk=8.6.11=h7bc2e8c_0
toolz=0.11.2=pyhd8ed1ab_0
tornado=6.1=py38h9ed2024_0
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.0=pyhd3eb1b0_0
typing-extensions=4.0.0=hd8ed1ab_0
typing_extensions=4.0.0=pyha770c72_0
tzdata=2021e=he74cb21_0
tzlocal=4.1=py38h50d1736_1
urllib3=1.26.7=pyhd8ed1ab_0
wbdata=0.3.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py38_0
wrapt=1.13.3=py38h96a0964_1
x264=1!161.3030=h0d85af4_1
xlrd=2.0.1=pyhd8ed1ab_3
xlsxwriter=3.0.2=pyhd8ed1ab_0
xlwt=1.3.0=pypi_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
yarl=1.7.2=py38h96a0964_1
zeromq=4.3.4=h23ab428_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
